services/analyzer-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiokafka==0.11.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
redis==5.0.8
elasticsearch==8.15.1
tenacity==9.0.0
prometheus-fastapi-instrumentator==7.0.0
python-jose==3.3.0
python-multipart==0.0.9
python-whois==0.9.4
dnspython==2.7.0
dkimpy==1.1.8
